Add the `enrich` operator
