+++ libtenzir/src/detail/loader_saver_resolver.cpp
@@ -8,97 +8,11 @@
 
 #include <tenzir/concept/parseable/tenzir/identifier.hpp>
 #include <tenzir/detail/loader_saver_resolver.hpp>
+#include <tenzir/prepend_token.hpp>
 
 namespace tenzir::detail {
 
 namespace {
-
-template <typename String>
-class prepend_token final : public parser_interface {
-public:
-  prepend_token(located<String> token, parser_interface& next)
-    : token_{std::move(token)}, next_{next} {
-  }
-  prepend_token(std::nullopt_t, parser_interface& next)
-    : token_{std::nullopt}, next_{next} {
-  }
-
-  auto accept_shell_arg() -> std::optional<located<std::string>> override {
-    if (token_) {
-      auto tmp = make_located_string();
-      token_ = std::nullopt;
-      return tmp;
-    }
-    return next_.accept_shell_arg();
-  }
-
-  auto peek_shell_arg() -> std::optional<located<std::string>> override {
-    if (token_) {
-      return make_located_string();
-    }
-    return next_.peek_shell_arg();
-  }
-
-  auto accept_identifier() -> std::optional<identifier> override {
-    TENZIR_ASSERT(not token_);
-    return next_.accept_identifier();
-  }
-
-  auto peek_identifier() -> std::optional<identifier> override {
-    TENZIR_ASSERT(not token_);
-    return next_.peek_identifier();
-  }
-
-  auto accept_equals() -> std::optional<location> override {
-    TENZIR_ASSERT(not token_);
-    return next_.accept_equals();
-  }
-
-  auto accept_char(char c) -> std::optional<location> override {
-    TENZIR_ASSERT(not token_);
-    return next_.accept_char(c);
-  }
-
-  auto parse_operator() -> located<operator_ptr> override {
-    TENZIR_ASSERT(not token_);
-    return next_.parse_operator();
-  }
-
-  auto parse_expression() -> tql::expression override {
-    TENZIR_ASSERT(not token_);
-    return next_.parse_expression();
-  }
-
-  auto parse_legacy_expression() -> located<expression> override {
-    TENZIR_ASSERT(not token_);
-    return next_.parse_legacy_expression();
-  }
-
-  auto parse_extractor() -> tql::extractor override {
-    TENZIR_ASSERT(not token_);
-    return next_.parse_extractor();
-  }
-
-  auto at_end() -> bool override {
-    return not token_ && next_.at_end();
-  }
-
-  auto current_span() -> location override {
-    if (token_) {
-      return token_->source;
-    }
-    return next_.current_span();
-  }
-
-private:
-  auto make_located_string() const -> located<std::string> {
-    TENZIR_ASSERT(token_);
-    return {std::string{token_->inner}, token_->source};
-  }
-
-  std::optional<located<String>> token_;
-  parser_interface& next_;
-};
 
 template <typename String>
 auto make_located_string_view(const located<String>& str, size_t pos = 0,
